don' mess with IPTables if they've already been patched.  Changes to be committed: 	modified:   InstallPortableOpenERP.sh 	modified:   ipoerpMakeOerpServerConfigFile.sh 	modified:   ipoerpPatchIPTables.sh

+++ ipoerpPatchIPTables.sh
@@ -12,11 +12,16 @@ fi
 #
 export SCRIPTNAME="odoo-${SITENAME}"
 #
+declare ALREADYPATCHED=$(cat /etc/default/iptables | grep -c "\-A INPUT.*dport ${ACCESS_PORT}.*ACCEPT")
 #
-echo "Creating /etc/default/iptables.patch"
-rm -f /etc/default/iptables.patch
-# .  .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   .
-cat <<PATCHEOF> /etc/default/iptables.patch
+if [[ 0 -lt ${ALREADYPATCHED} ]]
+then
+  echo "Port for OpenERP is already set to beopened."
+else
+  echo "Creating /etc/default/iptables.patch"
+  rm -f /etc/default/iptables.patch
+  # .  .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   .
+  cat <<PATCHEOF> /etc/default/iptables.patch
 --- /etc/default/iptables       2014-05-09 13:29:26.779041999 -0400
 +++ /etc/default/iX     2014-05-09 13:09:42.635041999 -0400
 @@ -35,6 +35,9 @@
@@ -31,12 +36,16 @@ cat <<PATCHEOF> /etc/default/iptables.patch
  -A INPUT -p tcp --dport 443 -j ACCEPT
 
 PATCHEOF
-# :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :
+  # :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :  :
+  #
+  echo "Patching /etc/default/iptables"
+  patch -u /etc/default/iptables /etc/default/iptables.patch
+  echo "Deleting patch."
+  rm -f /etc/default/iptables.patch
+  #
+  # cat /etc/default/iptables
 #
-patch -u /etc/default/iptables /etc/default/iptables.patch
-rm -f /etc/default/iptables.patch
-#
-cat /etc/default/iptables
+fi
 #
 echo "Trying upstart : ${SCRIPTNAME}"
 stop ${SCRIPTNAME}
